@@ -15,7 +15,7 @@
         }
     ],
     "require": {
-        "php": ">=7.2",
+        "php": ">7.4",
         "neomerx/cors-psr7": "^2.0",
         "psr/http-factory": "^1.0",
         "psr/http-server-middleware": "^1.0"
@@ -31,7 +31,7 @@
         }
     },
     "require-dev": {
-        "qlimix/code-standard": "^1.0"
+        "qlimix/code-standard": "^3.0"
     },
     "config": {
         "sort-packages": true
